Chore: dashboard refactorized, RAG assistant failed
