+++ frontend/src/components/dashboard/ProfileTab.jsx
@@ -1,0 +1,140 @@
+import { useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
+const ProfileTab = ({ userData, profileData, onProfileUpdate }) => {
+  const navigate = useNavigate()
+
+  const handleEditProfile = () => {
+    navigate('/edit-profile')
+  }
+
+  if (!userData || !profileData) {
+    return (
+      <div className="text-center py-8">
+        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-orange-500 mx-auto mb-4"></div>
+        <p className="text-gray-600">Cargando información del perfil...</p>
+      </div>
+    )
+  }
+
+  return (
+    <div>
+      <h2 className="text-2xl font-bold text-gray-800 mb-6">Mi Perfil</h2>
+      
+      <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
+        {/* Personal Info Card */}
+        <div className="bg-white border-2 border-gray-200 rounded-xl p-6 shadow-lg">
+          <h3 className="text-xl font-semibold text-gray-800 mb-4 flex items-center">
+            <svg className="w-5 h-5 text-orange-500 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
+            </svg>
+            Información Personal
+          </h3>
+          <div className="space-y-3">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Email</label>
+              <p className="text-gray-900">{userData.email}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Nombre Completo</label>
+              <p className="text-gray-900">{profileData.full_name || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Teléfono</label>
+              <p className="text-gray-900">{profileData.phone || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Ubicación</label>
+              <p className="text-gray-900">{profileData.location || 'No especificado'}</p>
+            </div>
+          </div>
+        </div>
+
+        {/* Professional Info Card */}
+        <div className="bg-white border-2 border-gray-200 rounded-xl p-6 shadow-lg">
+          <h3 className="text-xl font-semibold text-gray-800 mb-4 flex items-center">
+            <svg className="w-5 h-5 text-orange-500 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 13.255A23.931 23.931 0 0112 15c-3.183 0-6.22-.62-9-1.745M16 6V4a2 2 0 00-2-2h-4a2 2 0 00-2 2v2m4 6h.01M5 20h14a2 2 0 002-2V8a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z" />
+            </svg>
+            Información Profesional
+          </h3>
+          <div className="space-y-3">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Área de Interés</label>
+              <p className="text-gray-900">{profileData.area_of_interest || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Experiencia</label>
+              <p className="text-gray-900">{profileData.experience_level || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Habilidades</label>
+              <p className="text-gray-900">{profileData.skills || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Objetivos Profesionales</label>
+              <p className="text-gray-900">{profileData.professional_goals || 'No especificado'}</p>
+            </div>
+          </div>
+        </div>
+
+        {/* Education Card */}
+        <div className="bg-white border-2 border-gray-200 rounded-xl p-6 shadow-lg">
+          <h3 className="text-xl font-semibold text-gray-800 mb-4 flex items-center">
+            <svg className="w-5 h-5 text-orange-500 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 14l9-5-9-5-9 5 9 5z" />
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 14l6.16-3.422a12.083 12.083 0 01.665 6.479A11.952 11.952 0 0012 20.055a11.952 11.952 0 00-6.824-2.998 12.078 12.078 0 01.665-6.479L12 14z" />
+            </svg>
+            Formación
+          </h3>
+          <div className="space-y-3">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Educación</label>
+              <p className="text-gray-900">{profileData.education || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Certificaciones</label>
+              <p className="text-gray-900">{profileData.certifications || 'No especificado'}</p>
+            </div>
+          </div>
+        </div>
+
+        {/* Preferences Card */}
+        <div className="bg-white border-2 border-gray-200 rounded-xl p-6 shadow-lg">
+          <h3 className="text-xl font-semibold text-gray-800 mb-4 flex items-center">
+            <svg className="w-5 h-5 text-orange-500 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
+            </svg>
+            Preferencias
+          </h3>
+          <div className="space-y-3">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Modalidad de Trabajo</label>
+              <p className="text-gray-900">{profileData.work_modality || 'No especificado'}</p>
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700">Salario Esperado</label>
+              <p className="text-gray-900">{profileData.expected_salary || 'No especificado'}</p>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      {/* Edit Profile Button */}
+      <div className="mt-8 text-center">
+        <button
+          onClick={handleEditProfile}
+          className="bg-gradient-to-r from-orange-500 to-orange-300 text-white px-8 py-3 rounded-lg font-semibold hover:from-orange-600 hover:to-orange-400 transition duration-200 shadow-lg"
+        >
+          <svg className="w-5 h-5 inline-block mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z" />
+          </svg>
+          Editar Perfil
+        </button>
+      </div>
+    </div>
+  )
+}
+
+export default ProfileTab
